backend/requirements.txt
flask
requests
flask-sqlalchemy
psycopg2-binary
flask-login
werkzeug
